refactor: 타입 중복 제거 및 pino logger 추가
- type.ts 파일 제거, db/schema.ts의 타입 추론 사용
- db/index.ts에서 타입 re-export
- 단일 진실 공급원(Single Source of Truth) 확보
- pino logger 추가 (src/main/common/logger.ts)
- db.ts에 drizzle logger 연동 (개발 모드 SQL 쿼리 로깅)
