+++ apps/desktop/package.json
@@ -31,78 +31,80 @@
     "@dnd-kit/utilities": "^3.2.2",
     "@electron-toolkit/preload": "^3.0.2",
     "@electron-toolkit/utils": "^4.0.0",
-    "@fontsource/roboto-mono": "^5.2.6",
+    "@fontsource/roboto-mono": "^5.2.8",
     "@radix-ui/react-dialog": "^1.1.15",
-    "@radix-ui/react-dropdown-menu": "^2.1.15",
-    "@radix-ui/react-popover": "^1.1.14",
-    "@radix-ui/react-scroll-area": "^1.2.9",
+    "@radix-ui/react-dropdown-menu": "^2.1.16",
+    "@radix-ui/react-popover": "^1.1.15",
+    "@radix-ui/react-scroll-area": "^1.2.10",
     "@radix-ui/react-slot": "^1.2.3",
-    "@radix-ui/react-tooltip": "^1.2.7",
-    "@tailwindcss/vite": "^4.1.11",
-    "@tanstack/react-query": "^5.81.5",
+    "@radix-ui/react-tooltip": "^1.2.8",
+    "@tailwindcss/vite": "^4.1.16",
+    "@tanstack/react-query": "^5.90.5",
     "autosize": "^6.0.1",
-    "better-sqlite3": "^12.2.0",
+    "better-sqlite3": "^12.4.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "dotenv": "^17.0.1",
-    "drizzle-orm": "^0.44.2",
+    "dotenv": "^17.2.3",
+    "drizzle-orm": "^0.44.7",
     "electron-updater": "^6.6.2",
-    "es-toolkit": "^1.39.10",
-    "fs-extra": "^11.2.0",
-    "lucide-react": "^0.525.0",
+    "es-toolkit": "^1.41.0",
+    "fs-extra": "^11.3.2",
+    "lucide-react": "^0.548.0",
     "merge-refs": "^2.0.0",
     "next-themes": "^0.4.6",
-    "react-hotkeys-hook": "^5.1.0",
-    "react-router": "^7.6.3",
-    "sonner": "^1.7.4",
+    "pino": "^10.1.0",
+    "pino-pretty": "^13.1.2",
+    "react-hotkeys-hook": "^5.2.1",
+    "react-router": "^7.9.4",
+    "sonner": "^2.0.7",
     "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^4.1.11",
-    "tw-animate-css": "^1.3.5",
-    "uuid": "^11.1.0",
-    "zustand": "^5.0.6"
+    "tailwindcss": "^4.1.16",
+    "tw-animate-css": "^1.4.0",
+    "uuid": "^13.0.0",
+    "zustand": "^5.0.8"
   },
   "devDependencies": {
     "@electron-toolkit/eslint-config-prettier": "^3.0.0",
     "@electron-toolkit/eslint-config-ts": "^3.1.0",
-    "@electron-toolkit/tsconfig": "^1.0.1",
-    "@eslint/compat": "^1.3.1",
+    "@electron-toolkit/tsconfig": "^2.0.0",
+    "@eslint/compat": "^1.4.0",
     "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.30.1",
-    "@tanstack/eslint-plugin-query": "^5.81.2",
-    "@testing-library/dom": "^10.4.0",
+    "@eslint/js": "^9.38.0",
+    "@tanstack/eslint-plugin-query": "^5.91.2",
+    "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^16.3.0",
     "@types/autosize": "^4.0.3",
     "@types/better-sqlite3": "^7.6.13",
     "@types/fs-extra": "^11.0.4",
-    "@types/micromatch": "^4.0.9",
-    "@types/node": "^24.0.10",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
-    "@typescript-eslint/eslint-plugin": "^8.35.1",
-    "@typescript-eslint/rule-tester": "^8.35.1",
-    "@vitejs/plugin-react": "^4.6.0",
+    "@types/micromatch": "^4.0.10",
+    "@types/node": "^24.9.1",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.2",
+    "@typescript-eslint/eslint-plugin": "^8.46.2",
+    "@typescript-eslint/rule-tester": "^8.46.2",
+    "@vitejs/plugin-react": "^5.1.0",
     "@voyl/eslint-plugin-voyl": "workspace:*",
-    "drizzle-kit": "^0.31.4",
-    "electron": "37.2.0",
+    "drizzle-kit": "^0.31.5",
+    "electron": "38.4.0",
     "electron-builder": "26.0.12",
     "electron-devtools-installer": "^4.0.0",
-    "electron-vite": "3.1.0",
-    "eslint": "^9.30.1",
+    "electron-vite": "4.0.1",
+    "eslint": "^9.38.0",
     "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "happy-dom": "^18.0.1",
-    "lint-staged": "^16.1.2",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "happy-dom": "^20.0.8",
+    "lint-staged": "^16.2.6",
     "micromatch": "^4.0.8",
     "prettier": "^3.6.2",
-    "prettier-plugin-tailwindcss": "^0.6.13",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
+    "prettier-plugin-tailwindcss": "^0.7.1",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
     "tsconfig-paths": "^4.2.0",
     "typescript": "5.7.3",
-    "vite": "^6.3.5",
+    "vite": "^7.1.12",
     "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^3.2.4"
+    "vitest": "^4.0.3"
   }
 }
