+++ apps/desktop/src/main/repo/app/fs/index.ts
@@ -1,6 +1,7 @@
 import fse from 'fs-extra'
 import { PATH } from './const.js'
-import { App } from '../type.js'
+// eslint-disable-next-line voyl/same-level-import
+import type { App } from '../db/index.js'
 
 async function create({ version, workspaceDirPath }: App): Promise<void> {
   await fse.writeJson(PATH, {
