+++ apps/desktop/src/main/repo/app/index.ts
@@ -2,7 +2,7 @@ import AppDb from './db/index.js'
 import AppFs from './fs/index.js'
 // eslint-disable-next-line voyl/same-level-import
 import SyncMetadataRepo from '../syncMetadata/index.js'
-import { App } from './type.js'
+import type { App } from './db/index.js'
 
 async function exists(): Promise<boolean> {
   return await AppDb.existsTable()
