--- apps/desktop/src/main/repo/app/type.ts
@@ -1,4 +1,0 @@
-export interface App {
-  version: string
-  workspaceDirPath: string
-}
