--- apps/desktop/src/main/repo/node/type.ts
@@ -1,6 +1,0 @@
-export interface Node {
-  id: string
-  childIds: string[]
-  title: string
-  content: string
-}
