--- apps/desktop/src/main/repo/syncMetadata/type.ts
@@ -1,5 +1,0 @@
-export interface SyncMetadata {
-  path: string
-  tableName: string
-  syncedAt: number
-}
