--- apps/desktop/src/main/repo/workspace/type.ts
@@ -1,4 +1,0 @@
-export interface Workspace {
-  nodeTypes: string[]
-  attributes: string[]
-}
